With save_init_splits now write 3 sets of particles:
splitX_Npts_wo_init
splitX_Npts_w_init
splitX_Npts_w_opt

We used to have wo_opt and w_opt, but this was a misnomer as this was before and after the init phase.

The splits named w_opt will only be written when the multiscale option is used.