Fix #770 by adding an optional flag to computing the mean before warp.
It's the mean before warp computation step that requires the images be
the same size.  The result of this is not used by Studio, so I have
added a flag that disables this section of code.  It's unclear if this
section and potential nrrd output is really used elsewhere, but I have
left it as the default behavior so that the reconstruction executables
are unchanged.