Adding thread safe wrappers for Mesh reading.
Some VTK readers are not thread-safe, such as PLY.  Called at the same
time from multiple threads, they will cause crashes.  I was previously
handling this in Studio on its own, but now the Project class needs to
be able to inspect mesh file for scalars to determine what features
may be available.  As a result, code outside of Studio needs to be
able to safely read mesh files, so I'm moving this here.