Do not consider faces that result in NaN barycentric coordinates
This is a temporary fix until we either add a QC check (e.g. something
like vtkCleanPolyData) or make other changes.

The bug that this fixes is where the mesh contains degenerate faces
(such as two vertices of a triangle that land on the same location).
The current computation returns barycentric coordinates containing
NaNs which end up propagating to many other variables resulting in
either infinite loops or crashes.