Add support for file associations for MacOS (#1913).
I forgot how complicated it is for MacOS and finally discovered why I never really got it working with modal splashscreens in the past.  A modal dialog blocks the main event filter, so the QEvent::fileOpen must wait until the splashscreen goes away.  Very annoying.

The solution is to run QApplication::processEvents() a few times in main to get the QEvent::fileOpen to get processed before the splash screen displays.