Remove LD_LIBRARY_PATH, now I remember why I removed it before:
tar can't run with it set:

/bin/tar
tar: relocation error: /lib/x86_64-linux-gnu/libacl.so.1: symbol getxattr, version ATTR_1.0 not defined in file libattr.so.1 with link time reference

The problem is that I incorrectly copied the linuxdepoyqt binary to
/usr/bin in the docker image.  It has its own libQt5Core.so.5 that it
uses and it couldn't find it.  The solution is to symlink it instead.