Remove warning about unable to read features from mesh files.
I'm not sure what to do about this in general.  The problem is that the python use cases set the relative filenames with values like "../../meshes/foo.vtk", but they haven't saved the project yet, so the project class has no idea that it's path is going to be "Output/UseCase/shape_models", so there's no way for it to resolve these filenames until after it's been saved.