Reorder libraries to fix #1947
The problem in #1947 is that libMesh depends on other libs and one of
them ends up including xlnt before VTK gets included, so the symbol
"normal_nameMatchesAscii" from XLNT's expat gets used instead of VTK.

VTK tries to name mangle the symbols, but they clearly missed some.
It's tough because this expat library seems to be build off of crazy
preprocess macro magic, so I'm not surpised that the name mangling is
wrong.