Set Eigen threads to the same as TBB.
This came about due to a bug or unexpected interaction with calling DeepSSM commands from Studio.  After the image registration stuff runs, the OpenMP thread count is set to a crazy high level (e.g. 384 on a 16-core machine), which results in poor eigen performance per this page:

https://eigen.tuxfamily.org/dox/TopicMultiThreading.html

"Warning: On most OS it is very important to limit the number of threads to the number of physical cores, otherwise significant slowdowns are expected, especially for operations involving dense matrices."