DeepSSM input verification (#2196)
* Throw exception when number of original domains doesn't match.

* When images or original files are not present, disallow deepssm mode

* Throw useful exception when image size will be too small.

* Add timing to steps of prep.

* Add one more timer

* Set Eigen threads to the same as TBB.

This came about due to a bug or unexpected interaction with calling DeepSSM commands from Studio.  After the image registration stuff runs, the OpenMP thread count is set to a crazy high level (e.g. 384 on a 16-core machine), which results in poor eigen performance per this page:

https://eigen.tuxfamily.org/dox/TopicMultiThreading.html

"Warning: On most OS it is very important to limit the number of threads to the number of physical cores, otherwise significant slowdowns are expected, especially for operations involving dense matrices."

* Add Fine Tuning plot, combine csv

* Fix logger being closed for FT

* Allow decay LR to be off

* Fix formatting

* Reinitialze optimizer for fine tuning, use fine tuning learning rate

* Parse table doubles and limit precision for display.

* Fix table display (digits)

* Update parameter names and tooltips

* Reorganize DeepSSM Prep dialog a bit.
Change percent variability to be consistent
Add read-only training percent to show the user the amount that will be used.

* Update screenshot.

* Improve test mesh loading

* Fix TL-DeepSSM when Decay Learning is off.

* Replace shapeworks executable usage with Python API of MeshWarper.

* Fix typos

* Simplify and improve get_mesh_distance, also write back distance field to prediction mesh.

* Add more room for image spacing.

* Add ability to set reconstructed meshes directly.

* Shift test mesh distance calc and results into Python from Studio.

* Fix compile

* Fix image name for train/test when image is not already selected.

* Fix problem with boost create_directories