Add patch for OpenVDB with Clang17 based on this PR:
https://github.com/AcademySoftwareFoundation/openvdb/pull/1977/files

I'm doing the patch instead of upgrading because upgrading requires
forcing a new version of GCC which will cascade other dependency
upgrades.  We may end up doing them, but we need to do this one step
at a time and I need the CI/CD working in order to test things.